feat: add demo for shareToTimeline
