+++ example/package.json
@@ -19,7 +19,7 @@
     "@react-native-ohos/react-native-wechat-lib": "file:../react-native-ohos-react-native-wechat-lib-3.0.6-rc.3.tgz",
     "react": "18.2.0",
     "react-native": "0.72.5",
-    "react-native-harmony": "npm@react-native-oh/react-native-harmony@^0.72.67"
+    "react-native-harmony": "npm:@react-native-oh/react-native-harmony@^0.72.67"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
@@ -28,7 +28,7 @@
     "@react-native-community/eslint-config": "^3.2.0",
     "@react-native/eslint-config": "^0.74.0",
     "@react-native/metro-config": "^0.72.6",
-    "@rnoh/testerino": "npm@react-native-oh-tpc/testerino@0.0.9",
+    "@rnoh/testerino": "npm:@react-native-oh-tpl/testerino@0.0.9",
     "@tsconfig/react-native": "^2.0.2",
     "@types/chai": "^4.3.4",
     "@types/d3-scale-chromatic": "^3.0.0",
@@ -56,7 +56,7 @@
     "yargs": "^17.7.2"
   },
   "overrides": {
-    "@rnoh/react-native-harmony-cli": "npm@react-native-oh/react-native-harmony-cli@^0.0.28",
+    "@rnoh/react-native-harmony-cli": "npm:@react-native-oh/react-native-harmony-cli@^0.0.28",
     "@react-native-community/cli": "11.3.6",
     "@react-native/codegen": "0.74.0"
   },
